Extracted the new NewsScreen into a own file
